% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{artificial2sls_contaminated}
\alias{artificial2sls_contaminated}
\title{Artificial data set with outliers for illustration.}
\format{
A data frame with 100 observations (rows) and 15 variables (columns):
\describe{
\item{y}{dependent variable}
\item{x1}{intercept}
\item{x2}{relevant exogenous regressor}
\item{x3 - x10}{irrelevant exogenous regressors}
\item{x11}{relevant endogenous regressor}
\item{u}{structural error (in practice unobserved)}
\item{z11 - z12}{excluded instruments}
}
}
\usage{
artificial2sls_contaminated
}
\description{
A data set containing dependent variable, endogenous and exogenous
regressors, and excluded instruments for 2SLS models. The structural error is
also stored even though not observed in practice. Some errors are
contaminated, making these observations outliers.
}
\details{
The data frame has two additional attributes that store the indices
of the outliers, \code{"outliers"}, and their magnitudes
\code{"magnitude"}.
}
\keyword{datasets}
