% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{extract_regressors}
\alias{extract_regressors}
\title{Extract the first and second stage regressors of ivreg formula}
\usage{
extract_regressors(formula)
}
\arguments{
\item{formula}{A formula for the \link[ivreg:ivreg]{ivreg::ivreg} function, i.e. in format
\code{y ~ x1 + x2 | z1 + z2}.}
}
\value{
\code{extract_regressors} returns a list with three components:
\code{$yvar} stores the name of the dependent variable, \code{$first} the
names of the regressors of the first stage and \code{$second} of the second
stage regressors.
}
\description{
\code{extract_regressors} takes a formula object for \code{\link[ivreg:ivreg]{ivreg::ivreg()}}, i.e.
in a format of \code{y ~ x1 + x2 | x1 + z2} and extracts the different
elements in a list.
}
