% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{factory_indicators}
\alias{factory_indicators}
\title{Function factory for creating indicators from their names}
\usage{
factory_indicators(n)
}
\arguments{
\item{n}{An integer specifying the length of the indicators.}
}
\value{
\code{factory_indicators} returns a function called \code{creator()}.
}
\description{
\code{factory_indicators} creates a function that takes the name of an
indicator and returns the corresponding indicator to be used in a regression.
For user-specified indicators, it extracts the corresponding column from the
uis matrix.
}
\details{
Argument \code{n} should equal the number of observations in the
data set which will be augmented with the indicators.

The created function takes a name of an indicator and the original uis
argument that was used in indicator saturation and returns the indicator.
}
