% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{twosls.alt}
\alias{twosls.alt}
\title{2SLS estimator alternative}
\usage{
twosls.alt(formula, data)
}
\arguments{
\item{formula}{A formula in the format \code{y ~ x1 + x2 | z1 + z2}.}

\item{data}{A data frame with all necessary variables y, x, and z.}
}
\description{
Test whether is faster than \code{\link[=twosls]{twosls()}} but this does not seem to be
the case.
}
\section{WARNING}{

The return value is given class \code{\link[ivreg:ivreg]{ivreg::ivreg()}} but it is not a true
\code{"ivreg"} object. This does not pose any problems for internal use but
should not be used outside of its usage in its current form. The class
assignment is likely to change in the future.
}

\keyword{internal}
