% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{twosls.fit}
\alias{twosls.fit}
\title{2SLS estimator}
\usage{
twosls.fit(x, y, z)
}
\description{
Stripped down version of \code{\link[ivreg:ivreg.fit]{ivreg::ivreg.fit()}} that does not allow for
weights, offset, other methods than 2SLS, and does not calculate influence
statistics.
}
