% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{artificial2sls_contaminated}
\alias{artificial2sls_contaminated}
\title{Artificial data set with outliers for illustration.}
\format{
A data frame with 100 observations (rows) and 16 variables (columns):
\tabular{rl}{
\strong{name} \tab \strong{variable description} \cr
y \tab dependent variable \cr
x1 \tab intercept \cr
x2 \tab relevant exogenous regressor \cr
x3 \tab irrelevant exogenous regressor \cr
x4 \tab irrelevant exogenous regressor \cr
x5 \tab irrelevant exogenous regressor \cr
x6 \tab irrelevant exogenous regressor \cr
x7 \tab irrelevant exogenous regressor \cr
x8 \tab irrelevant exogenous regressor \cr
x9 \tab irrelevant exogenous regressor \cr
x10 \tab irrelevant exogenous regressor \cr
x11 \tab relevant endogenous regressor \cr
u \tab structural error (in practice unobserved) \cr
z11 \tab excluded instrument \cr
z12 \tab excluded instrument \cr
id \tab unique observation identifier
}
}
\usage{
artificial2sls_contaminated
}
\description{
A data set containing dependent variable, endogenous and exogenous
regressors, and excluded instruments for 2SLS models. The structural error is
also stored even though not observed in practice. Some errors are
contaminated, making these observations outliers.
}
\details{
The data frame has two additional attributes that store the indices
of the outliers, \code{"outliers"}, and their magnitudes
\code{"magnitude"}.
}
\keyword{datasets}
