% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{artificial2sls_shiny}
\alias{artificial2sls_shiny}
\title{Artificial data set without outliers prepared for shiny application.}
\format{
A data frame with 100 observations (rows) and 17 variables (columns):
\tabular{rl}{
\strong{name} \tab \strong{variable description} \cr
y \tab dependent variable \cr
x1 \tab intercept \cr
x2 \tab relevant exogenous regressor \cr
x3 \tab irrelevant exogenous regressor \cr
x4 \tab irrelevant exogenous regressor \cr
x5 \tab irrelevant exogenous regressor \cr
x6 \tab irrelevant exogenous regressor \cr
x7 \tab irrelevant exogenous regressor \cr
x8 \tab irrelevant exogenous regressor \cr
x9 \tab irrelevant exogenous regressor \cr
x10 \tab irrelevant exogenous regressor \cr
x11 \tab relevant endogenous regressor \cr
u \tab structural error (in practice unobserved) \cr
z11 \tab excluded instrument \cr
z12 \tab excluded instrument \cr
id \tab unique observation identifier \cr
is.outlier \tab factor variable whether the observation is an outlier (\code{1}) or not (\code{0})
}
}
\usage{
artificial2sls_shiny
}
\description{
Artificial data set without outliers prepared for shiny application.
}
\keyword{datasets}
