% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{artificial2sls_shiny}
\alias{artificial2sls_shiny}
\title{Artificial data set prepared for shiny application}
\format{
A data frame with 100 observations (rows) and 17 variables (columns)
}
\usage{
artificial2sls_shiny
}
\description{
Similar to "artificial2sls" with a few minor additions.
}
\keyword{datasets}
