% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivgets-package.R
\docType{package}
\name{ivgets-package}
\alias{ivgets-package}
\title{ivgets: General to Specific Modeling and Indicator Saturation in 2SLS Models}
\description{
Provides facilities of general to specific model selection for exogenous regressors in 2SLS models. Furthermore, indicator saturation methods can be used to detect outliers and structural breaks in the sample.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jkurle/ivgets}
  \item Report bugs at \url{https://github.com/jkurle/ivgets/issues}
}

}
\author{
\strong{Maintainer}: Kurle Jonas \email{mail@jonaskurle.com} (\href{https://orcid.org/0000-0003-2197-2012}{ORCID})

}
\keyword{internal}
