% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivgets.R
\docType{package}
\name{ivgets-package}
\alias{ivgets-package}
\title{ivgets: A package for general to specific modeling and indicator saturation in 2SLS models}
\description{
Provides facilities of general to specific model selection for exogenous
regressors in 2SLS models. Furthermore, indicator saturation methods can be
used to detect outliers and structural breaks in the sample.
}
