% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getsiv.R
\name{ivgets}
\alias{ivgets}
\title{Gets modelling for 2SLS models}
\usage{
ivgets(
  formula,
  data,
  gum.result = NULL,
  t.pval = 0.05,
  wald.pval = t.pval,
  do.pet = TRUE,
  ar.LjungB = NULL,
  arch.LjungB = NULL,
  normality.JarqueB = NULL,
  include.gum = FALSE,
  include.1cut = FALSE,
  include.empty = FALSE,
  max.paths = NULL,
  turbo = FALSE,
  tol = 1e-07,
  max.regs = NULL,
  print.searchinfo = TRUE,
  alarm = FALSE,
  keep_exog = NULL,
  overid = NULL,
  weak = NULL
)
}
\arguments{
\item{keep_exog}{A numeric vector of indices or a character vector of names
corresponding to the exogenous regressors in the \code{data} that should not
be selected over. Default \code{NULL} means that selection is over all
exogenous regressors. If an intercept has been specified in the
\code{formula} but is not already included in the \code{data}, then it can be
kept by either including the index \code{0} or the character
\code{"(Intercept)"}, respectively, as an element in \code{keep_exog}.}
}
\description{
Gets modelling for 2SLS models
}
