% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{twosls}
\alias{twosls}
\title{2SLS estimator wrapper}
\usage{
twosls(formula, data)
}
\description{
Stripped down version of \code{\link[ivreg:ivreg]{ivreg::ivreg()}} that does not allow for
weights, offset, other methods than 2SLS, and does not calculate influence
statistics.
}
\details{
Based on \code{\link[=twosls.fit]{twosls.fit()}}.
}
